+++ lib/WorseReflection/Core/Type/GenericClassType.php
@@ -14,11 +14,19 @@ use Phpactor\WorseReflection\Core\Types;
 class GenericClassType extends ReflectedClassType implements IterableType, ClassLikeType
 {
     /**
+     * @var Type[]
+     */
+    protected array $arguments;
+
+    /**
      * @param Type[] $arguments
      */
-    public function __construct(ClassReflector $reflector, ClassName $name, public array $arguments)
+    public function __construct(ClassReflector $reflector, ClassName $name, array $arguments)
     {
         parent::__construct($reflector, $name);
+        $this->reflector = $reflector;
+        $this->name = $name;
+        $this->arguments = array_values($arguments);
     }
 
     public function __toString(): string
@@ -50,11 +58,31 @@ class GenericClassType extends ReflectedClassType implements IterableType, Class
 
     public function accepts(Type $type): Trinary
     {
-        if ($this->is($type)->isTrue()) {
-            return Trinary::true();
+        if (!$type instanceof GenericClassType) {
+            return parent::accepts($type);
         }
 
-        return Trinary::false();
+        if (!parent::accepts($type)->isTrue()) {
+            return Trinary::false();
+        }
+
+        $typeArguments = $type->arguments;
+
+        // horrible hack for "special" types which have > 1 "constructors"
+        if (in_array($type->name()->__toString(), IterableTypeResolver::iterableClasses())) {
+            array_unshift($typeArguments, TypeFactory::arrayKey());
+        }
+
+        foreach ($this->arguments as $index => $argument) {
+            if (!isset($typeArguments[$index])) {
+                return Trinary::false();
+            }
+            if (!$argument->accepts($typeArguments[$index])->isTrue()) {
+                return Trinary::false();
+            }
+        }
+
+        return Trinary::true();
     }
 
     public function replaceArgument(int $offset, Type $type): self