+++ lib/WorseReflection/Core/Type/ReflectedClassType.php
@@ -65,6 +65,15 @@ class ReflectedClassType extends ClassType
             return Trinary::true();
         }
 
+        if ($type instanceof UnionType) {
+            foreach ($type->types as $uType) {
+                if (!$this->accepts($uType)->isTrue()) {
+                    return Trinary::false();
+                }
+            }
+            return Trinary::true();
+        }
+
         if (!$type instanceof ClassType) {
             return Trinary::false();
         }
@@ -105,7 +114,7 @@ class ReflectedClassType extends ClassType
     {
         try {
             return $this->reflector->reflectClassLike($this->name());
-        } catch (NotFound) {
+        } catch (NotFound $e) {
         }
         return null;
     }